# sql: use a single column family for the KV* benchmarks