@petermattis 2016-06-25
I'm not sure why the `KVDelete` benchmarks did not improve. Delete needs some optimization love as it has shown to be a performance outlier on other benchmarks. Something to investigate in the future.