@tschottdorf 2016-06-25
Am I reading this right that KV is slower for small inserts, then better for larger ones - or is it the other way around? What's going on there?

```
KVInsert1-32       642µs ± 1%   481µs ± 2%  -25.16%  (p=0.000 n=20+19)
KVInsert10-32      801µs ± 1%   661µs ± 2%  -17.46%  (p=0.000 n=20+19)
KVInsert100-32    2.03ms ± 2%  2.03ms ± 1%     ~     (p=0.602 n=20+20)
KVInsert1000-32   12.0ms ± 2%  14.2ms ± 3%  +18.42%  (p=0.000 n=20+19)
KVInsert10000-32   124ms ± 8%   150ms ± 3%  +20.24%  (p=0.000 n=20+18)
```