@petermattis 2016-06-25
@tschottdorf Yes, you're reading that correctly. I think what we're seeing is that pgwire is faster than gRPC, but there is still overhead in the SQL level when processing large numbers of rows.